Create xyc_scatter_selection.py
This snippet shows a simple example of a Plotly scatterplot where the marker color is set by a third parameter which can be selected from a list. Two tags of the loaded view serve as x and y values, while the other tags of the view can be selected as color values.